requirements.txt
fastapi==0.115.12
uvicorn
gunicorn
timescaledb
pydantic
sqlalchemy
sqlmodel
python-dotenv
